NIFI-9437 Flowfile Expiration cannot exceed 24 days

+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-api/src/main/java/org/apache/nifi/web/search/attributematchers/ExpirationMatcher.java
@@ -35,7 +35,7 @@ public class ExpirationMatcher implements AttributeMatcher<Connection> {
     @Override
     public void match(final Connection component, final SearchQuery query, final List<String> matches) {
         if (containsKeyword(query)) {
-            final int expirationMillis = component.getFlowFileQueue().getFlowFileExpiration(TimeUnit.MILLISECONDS);
+            final long expirationMillis = component.getFlowFileQueue().getFlowFileExpiration(TimeUnit.MILLISECONDS);
 
             if (expirationMillis > 0) {
                 matches.add(MATCH_PREFIX + component.getFlowFileQueue().getFlowFileExpiration());
